Add download audioNotes

+++ src/store/apps/audioNotesActions.js
@@ -1,51 +1,54 @@
-import moment from 'moment';
-import httpService from '../services/httpService';
-import { getCurrentUser } from '../services/authService';
-import * as actions from './audioNotes';
-import config from '../../config.json';
-import { toast } from 'react-toastify';
+import moment from 'moment'
+import httpService from '../services/httpService'
+import { getCurrentUser } from '../services/authService'
+import * as actions from './audioNotes'
+import config from '../../config.json'
+import { toast } from 'react-toastify'
 
-const apiEndPoint = '/audioNotes';
-let userid;
-const user = getCurrentUser();
-const loadingInterval = Number(config.loadingInterval);
+const apiEndPoint = '/audioNotes'
+let userid
+const user = getCurrentUser()
+const loadingInterval = Number(config.loadingInterval)
 
-if (user) userid = user._id;
+if (user) userid = user._id
 
 export const loadAudioNotes = subjectId => async (dispatch, getState) => {
-  const { lastFetch } = getState().apps.audioNotes;
-  const diffInMinutes = moment().diff(moment(lastFetch), 'minutes');
-  if (diffInMinutes < loadingInterval) return;
-  const options = { headers: { userid } };
+  const { lastFetch } = getState().apps.audioNotes
+  const diffInMinutes = moment().diff(moment(lastFetch), 'minutes')
+  if (diffInMinutes < loadingInterval) return
+  const options = { headers: { userid } }
 
   try {
-    dispatch(actions.REQUEST_AUDIO_NOTES());
+    dispatch(actions.REQUEST_AUDIO_NOTES())
 
     const { data } = await httpService.get(
       `${apiEndPoint}/${subjectId}`,
       options
-    );
+    )
 
-    dispatch(actions.GET_AUDIO_NOTES(data));
+    dispatch(actions.GET_AUDIO_NOTES(data))
   } catch (error) {
-    console.log(error);
-    dispatch(actions.REQUEST_AUDIO_NOTES_FAIL());
+    console.log(error)
+    dispatch(actions.REQUEST_AUDIO_NOTES_FAIL())
   }
-};
+}
 
 export const createAudioNotesGroup = group => async dispatch => {
   try {
-    const { data } = await httpService.post(`${apiEndPoint}/groups`, group);
+    const { data } = await httpService.post(`${apiEndPoint}/groups`, group)
 
-    dispatch(actions.CREATE_AUDIO_NOTES_GROUP(data));
+    dispatch(actions.CREATE_AUDIO_NOTES_GROUP(data))
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
 
 export const updateAudioNotesGroup = (id, update) => async dispatch => {
   try {
-    const { data } = await httpService.patch(`${apiEndPoint}/groups/${id}`, update)
+    const { data } = await httpService.patch(
+      `${apiEndPoint}/groups/${id}`,
+      update
+    )
 
     dispatch(actions.UPDATE_AUDIO_NOTES_GROUP(data))
   } catch (error) {
@@ -55,45 +58,61 @@ export const updateAudioNotesGroup = (id, update) => async dispatch => {
 
 export const createAudioNote = (formData, params) => async dispatch => {
   try {
-    const options = { params: params };
-    const { data } = await httpService.post(apiEndPoint, formData, options);
+    const options = { params: params }
+    const { data } = await httpService.post(apiEndPoint, formData, options)
 
-    dispatch(actions.CREATE_AUDIO_NOTE(data));
+    dispatch(actions.CREATE_AUDIO_NOTE(data))
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
 
 export const patchAudioNote = (id, update) => async dispatch => {
   try {
-    const { data } = await httpService.patch(`${apiEndPoint}/${id}`, update);
+    const { data } = await httpService.patch(`${apiEndPoint}/${id}`, update)
 
-    dispatch(actions.UPDATE_AUDIO_NOTE({ data, update }));
+    dispatch(actions.UPDATE_AUDIO_NOTE({ data, update }))
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
 
 export const deleteAudioNoteGroup = group => async dispatch => {
   if (!group.children.length) {
     try {
-      await httpService.delete(`${apiEndPoint}/group/${group._id}`);
+      await httpService.delete(`${apiEndPoint}/group/${group._id}`)
 
-      dispatch(actions.DELETE_AUDIO_NOTE_GROUP(group._id));
+      dispatch(actions.DELETE_AUDIO_NOTE_GROUP(group._id))
     } catch (error) {
-      console.log(error);
+      console.log(error)
     }
   } else {
-    toast.error('Group must be empty to be deleted');
+    toast.error('Group must be empty to be deleted')
   }
-};
+}
 
 export const deleteAudioNote = audioNote => async dispatch => {
   try {
-    await httpService.delete(`${apiEndPoint}/${audioNote._id}`);
+    await httpService.delete(`${apiEndPoint}/${audioNote._id}`)
 
-    dispatch(actions.DELETE_AUDIO_NOTE(audioNote));
+    dispatch(actions.DELETE_AUDIO_NOTE(audioNote))
   } catch (error) {
-    console.log(error);
+    console.log(error)
   }
-};
+}
+
+export const deleteAudioNoteGroupAllNotes = group => async dispatch => {
+  if (group.children.length) {
+    for (const audioNote in group.children) {
+      try {
+        await httpService.delete(`${apiEndPoint}/${audioNote._id}`)
+
+        dispatch(actions.DELETE_AUDIO_NOTE(audioNote))
+      } catch (error) {
+        console.log(error)
+      }
+    }
+  } else {
+    toast.info('Group is already empty')
+  }
+}
